crawl/requirements.txt
requests==2.31.0
urllib3==2.0.7
json5==0.9.14
fake-useragent==1.4.0
mysql-connector-python==8.2.0
